requirements.txt
azure-core==1.25.1
azure-storage-blob==12.13.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
cx-Oracle==8.3.0
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
greenlet==1.1.3
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msrest==0.7.1
mysqlclient==2.1.1
numpy==1.23.2
oauthlib==3.2.1
openpyxl==3.0.10
pandas==1.4.4
pycparser==2.21
pymongo==4.2.0
PyMySQL==1.0.2
pyodbc==4.0.34
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.41
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wfastcgi==3.0.0
xlrd==2.0.1
